+++ packages/widgets/lib/src/elements/widgets/container.dart
@@ -59,5 +59,6 @@ class OpenWContainer extends Container {
             shadow: shadows,
             borders: borders,
           ),
+          clipBehavior: Clip.none,
         );
 }